+++ src/libopenrave/kinbody.cpp
@@ -5944,6 +5944,7 @@ void KinBody::Clone(InterfaceBaseConstPtr preference, int cloningoptions)
                 //BOOST_ASSERT(pgrabbedbody->GetName() == pbodyref->GetName());
 
                 GrabbedPtr pgrabbed(new Grabbed(pgrabbedbody,_veclinks.at(KinBody::LinkPtr(pgrabbedref->_pGrabbingLink)->GetIndex())));
+                pgrabbed->_grippername = pgrabbedref->_grippername;
                 pgrabbed->_tRelative = pgrabbedref->_tRelative;
                 pgrabbed->_setGrabberLinkIndicesToIgnore = pgrabbedref->_setGrabberLinkIndicesToIgnore; // can do this since link indices are the same
                 CopyRapidJsonDoc(pgrabbedref->_rGrabbedUserData, pgrabbed->_rGrabbedUserData);